lab1/requirements.txt
black==23.1.0
click==8.1.3
greenlet==2.0.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
PyMySQL==1.0.2
tk==0.1.0
tomli==2.0.1
typing_extensions==4.5.0

